Ditch the now-unused js_generic_base separation
