--- src/apps/js_generic/js_generic_base.h
@@ -1,13 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the Apache 2.0 License.
-#pragma once
-
-#include "ccf/app_interface.h"
-
-#include <memory>
-
-namespace ccf
-{
-  std::unique_ptr<ccf::endpoints::EndpointRegistry> make_user_endpoints_impl(
-    ccf::AbstractNodeContext& context);
-}
